+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
-## [3.2.2] - 2024-06-10
+### Added
+
+- Exposes route to parse incoming configuration
 
 ## [3.2.2] - 2024-06-10